# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 供应链金融系统 API文档
  description: 基于FISCO BCOS区块链的供应链金融系统接口文档
  contact:
    name: 开发团队
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Inferred Url
...
paths:
  /api/audit/entity/{entityType}/{entityId}:
    get:
      tags:
      - 审计日志管理
      summary: 查询实体的操作历史
      description: 查询特定实体的所有操作记录
      operationId: getEntityHistoryUsingGET
      parameters:
      - name: entityType
        in: path
        description: 实体类型
        required: true
        style: simple
        schema:
          type: string
      - name: entityId
        in: path
        description: 实体ID
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«审计日志»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/audit/logs:
    get:
      tags:
      - 审计日志管理
      summary: 分页查询审计日志
      description: 支持多条件组合查询审计日志
      operationId: queryAuditLogsUsingGET
      parameters:
      - name: actionType
        in: query
        description: 操作类型（CREATE, UPDATE, DELETE等）
        required: false
        style: form
        schema:
          type: string
      - name: endDate
        in: query
        description: 结束时间
        required: false
        style: form
        schema:
          type: string
          format: date-time
      - name: entityId
        in: query
        description: 实体ID
        required: false
        style: form
        schema:
          type: string
      - name: entityType
        in: query
        description: 实体类型
        required: false
        style: form
        schema:
          type: string
      - name: isSuccess
        in: query
        description: 是否成功
        required: false
        style: form
        schema:
          type: boolean
      - name: module
        in: query
        description: 操作模块（BILL, RECEIVABLE, WAREHOUSE_RECEIPT等）
        required: false
        style: form
        schema:
          type: string
      - name: page
        in: query
        description: 页码（从0开始）
        required: false
        style: form
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        description: 每页大小
        required: false
        style: form
        schema:
          type: integer
          format: int32
      - name: sortDirection
        in: query
        description: 排序方向（ASC, DESC）
        required: false
        style: form
        schema:
          type: string
      - name: sortField
        in: query
        description: 排序字段
        required: false
        style: form
        schema:
          type: string
      - name: startDate
        in: query
        description: 开始时间
        required: false
        style: form
        schema:
          type: string
          format: date-time
      - name: userAddress
        in: query
        description: 操作人地址
        required: false
        style: form
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«Page«审计日志»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/audit/logs/recent:
    get:
      tags:
      - 审计日志管理
      summary: 获取最近的审计日志
      description: 获取最近的操作记录
      operationId: getRecentLogsUsingGET
      parameters:
      - name: limit
        in: query
        description: 返回数量限制
        required: false
        style: form
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«审计日志»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/audit/logs/{id}:
    get:
      tags:
      - 审计日志管理
      summary: 获取审计日志详情
      description: 根据ID查询审计日志的详细信息
      operationId: getAuditLogByIdUsingGET
      parameters:
      - name: id
        in: path
        description: 审计日志ID
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«审计日志»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/audit/statistics:
    get:
      tags:
      - 审计日志管理
      summary: 统计审计日志
      description: 获取操作统计数据
      operationId: getStatisticsUsingGET
      parameters:
      - name: startDate
        in: query
        description: 开始时间
        required: false
        style: form
        schema:
          type: string
          format: date-time
      - name: endDate
        in: query
        description: 结束时间
        required: false
        style: form
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«审计日志统计»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/audit/tx/{txHash}:
    get:
      tags:
      - 审计日志管理
      summary: 根据交易哈希查询
      description: 根据区块链交易哈希查询相关的审计日志
      operationId: getLogsByTxHashUsingGET
      parameters:
      - name: txHash
        in: path
        description: 交易哈希
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«审计日志»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/audit/user/{userAddress}:
    get:
      tags:
      - 审计日志管理
      summary: 查询用户的操作历史
      description: 查询特定用户的所有操作记录
      operationId: getUserHistoryUsingGET
      parameters:
      - name: userAddress
        in: path
        description: 用户地址
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result«List«审计日志»»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
components:
  schemas:
...
    Pageable:
      title: Pageable
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: '#/components/schemas/Sort'
        unpaged:
          type: boolean
...
    Page«审计日志»:
      title: Page«审计日志»
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/审计日志'
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        size:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    Result«List«审计日志»»:
      title: Result«List«审计日志»»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/审计日志'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«Page«审计日志»»:
      title: Result«Page«审计日志»»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/Page«审计日志»'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Result«审计日志»:
      title: Result«审计日志»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/审计日志'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
    Result«审计日志统计»:
      title: Result«审计日志统计»
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: '#/components/schemas/审计日志统计'
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    Sort:
      title: Sort
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    UserOperationStats:
      title: UserOperationStats
      type: object
      properties:
        operationCount:
          type: integer
          description: 操作次数
          format: int64
        userAddress:
          type: string
          description: 用户地址
        userName:
          type: string
          description: 用户姓名
...
    审计日志:
      title: 审计日志
      required:
      - actionDesc
      - actionType
      - createdAt
      - isSuccess
      - module
      type: object
      properties:
        actionDesc:
          type: string
          description: 操作描述
          example: 创建票据
        actionType:
          type: string
          description: 操作类型
          example: CREATE
        changedFields:
          type: string
          description: 变更的字段
        createdAt:
          type: string
          description: 创建时间
          format: date-time
        duration:
          type: integer
          description: 操作时长（毫秒）
          format: int64
          example: 150
        entityId:
          type: string
          description: 实体ID
          example: bill001
        entityType:
          type: string
          description: 实体类型
          example: Bill
        errorMessage:
          type: string
          description: 错误信息
        isSuccess:
          type: boolean
          description: 操作是否成功
          example: true
        module:
          type: string
          description: 操作模块
          example: BILL
        newValue:
          type: string
          description: 操作后的数据
        oldValue:
          type: string
          description: 操作前的数据
        requestIp:
          type: string
          description: 请求IP地址
          example: 192.168.1.100
        requestMethod:
          type: string
          description: 请求方法
          example: POST
        requestUrl:
          type: string
          description: 请求URL
          example: /api/bill/create
        result:
          type: string
          description: 操作结果
          example: SUCCESS
        tenantId:
          type: string
          description: 租户ID
          example: tenant001
        txHash:
          type: string
          description: 区块链交易哈希
          example: '0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890'
        userAddress:
          type: string
          description: 操作人地址
          example: '0x1234567890abcdef1234567890abcdef12345678'
        userAgent:
          type: string
          description: 用户代理
          example: Mozilla/5.0...
        userName:
          type: string
          description: 操作人姓名
          example: 张三
      description: 系统操作审计日志实体
    审计日志统计:
      title: 审计日志统计
      type: object
      properties:
        actionTypeStats:
          type: object
          additionalProperties:
            type: integer
            format: int64
          description: 各操作类型统计
        averageDuration:
          type: integer
          description: 平均操作时长（毫秒）
          format: int64
        endTime:
          type: string
          description: 统计结束时间
          format: date-time
        failureCount:
          type: integer
          description: 失败操作次数
          format: int64
        moduleStats:
          type: object
          additionalProperties:
            type: integer
            format: int64
          description: 各模块操作统计
        startTime:
          type: string
          description: 统计开始时间
          format: date-time
        successCount:
          type: integer
          description: 成功操作次数
          format: int64
        successRate:
          type: number
          description: 成功率（%）
          format: double
        topUsers:
          type: array
          description: 用户操作排名（前10）
          items:
            $ref: '#/components/schemas/UserOperationStats'
        totalOperations:
          type: integer
          description: 总操作次数
          format: int64
      description: 审计日志的统计信息
...
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header